+++ noodles-vcf/examples/vcf_reheader.rs
@@ -12,11 +12,11 @@ use std::{
 use noodles_vcf as vcf;
 
 fn add_comment(header: &mut vcf::Header) {
-    use vcf::header::record::{value, Key};
+    use vcf::header::record::{Key, Value};
 
     header.insert(
         Key::other("comment").unwrap(),
-        value::Other::from("a comment added by noodles-vcf"),
+        Value::from("a comment added by noodles-vcf"),
     );
 }